+++ app/utils/callback_functions.py
@@ -20,14 +20,10 @@ from skimage import exposure
 import pandas as pd
 from pathlib import Path
 import os
-import subprocess
-import time
-import shlex
-import re
 
 ########################################################################
 ####                                                                ####
-####                             LAYOUT                             ####
+####                          Functions                             ####
 ####                                                                ####
 ########################################################################
 
@@ -409,1063 +405,6 @@ def convert_tiff_to_tif(input_file, output_file):
     with Image.open(input_file) as img:
         # Save the image as .TIF
         img.save(output_file)
-
-########################################################################
-####                                                                ####
-####                   preview.py functions                         ####
-####                                                                ####
-########################################################################
-
-def preamble_to_run_wrmXpress_preview(
-       platename,
-       volume, 
-       wells 
-):
-    """
-    This function prepares the wrmXpress command, output preview log file, command message, and first well.
-    ===============================================================================
-    Arguments:
-        - platename : str : Name of the plate
-        - volume : str : Path to the volume
-        - wells : list : List of well names
-    ===============================================================================
-    Returns:
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - output_preview_log_file : str : Path to the output preview log file
-        - command_message : str : A command message
-        - first_well : str : The first well in the list of wells
-    ===============================================================================
-    
-    """
-     # defining the yaml file path (same as the filepath from configure.py)
-    preview_yaml_platename = '.' + platename + '.yml'
-    preview_yaml_platenmaefull_yaml = Path(volume, preview_yaml_platename)
-    full_yaml = Path(volume, platename + '.yml')
-
-    update_yaml_file(
-                full_yaml, 
-                preview_yaml_platenmaefull_yaml,
-                {'wells': ['All']}
-            )
-            
-    if wells == 'All':
-            first_well = "A01"
-    else:
-            first_well = wells[0]
-
-
-        # Checking if input folder exists, and if not, create it,
-        # then subsequently copy the images into this folder
-        # Input and platename input folder paths
-        # necessary file paths
-    img_dir = Path(volume, platename)
-    input_dir = Path(volume, 'input')
-    platename_input_dir = Path(input_dir, platename)
-    plate_base = platename.split("_", 1)[0]
-    htd_file = Path(img_dir, f'{plate_base}.HTD')
-
-    # Clean and create directories
-    clean_and_create_directories(
-                input_path=Path(volume, 'input', platename),
-                work_path=Path(volume, 'work', platename)
-    )
-            
-            # Copy files to input directory
-    copy_files_to_input_directory(
-                platename_input_dir=platename_input_dir,
-                htd_file=htd_file,
-                img_dir=img_dir,    
-                plate_base=plate_base,
-                wells = first_well,
-                platename=platename
-            )
-
-        # Command message
-    command_message = f"```python wrmXpress/wrapper.py {platename}.yml {platename}```"
-
-    wrmxpress_command = f'python wrmXpress/wrapper.py {volume}/.{platename}.yml {platename}'
-    wrmxpress_command_split = shlex.split(wrmxpress_command)
-    output_preview_log_file = Path(volume, 'input', platename, f'{platename}_preview.log')
-    return wrmxpress_command_split, output_preview_log_file, command_message, first_well
-
-def motility_segment_fecundity_preview(volume, platename, wells, selection):
-    """
-    This function is used to preview the analysis of the selected options from
-    the configuration page, including motility, segment, and fecundity. This function
-    will run wrmXpress and return the path to the image, the figure, and the open status
-    if the first well has not already been analyzed.
-    ===============================================================================
-    Arguments:
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - wells : list : List of well names
-        - selection : str : Selection
-    ===============================================================================
-    Returns:
-        - path : str : Path to the image
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-        - open_status : bool : Open status of the alerts
-    ===============================================================================
-    """
-    # Check to see if first well already exists, if it does insert the img
-    # rather than running wrmXpress again
-    first_well_path = Path(volume, 'work', f'{platename}/{wells[0]}/img/{platename}_{wells[0]}.png')
-
-    # Check if the first well path exists
-    if os.path.exists(first_well_path):
-        # Checking the selection and changing the scale accordingly
-        if selection == 'motility':
-            scale = 'inferno'
-        else:
-            scale = 'gray'
-
-        # Open the image and create a figure
-        fig = create_figure_from_filepath(first_well_path, scale=scale)
-
-        # Return the path and the figure and the open status of the alerts
-        return f"```{first_well_path}```", fig, False, f'', False
-
-    wrmxpress_command_split, output_preview_log_file, command_message, first_well = preamble_to_run_wrmXpress_preview(
-        platename=platename, volume=volume, wells=wells
-    )
-    with open(output_preview_log_file, 'w') as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-
-        print('Running wrmXpress.')
-        docker_output.append('Running wrmXpress.')
-
-        while not os.path.exists(Path(volume, 'input')):
-            time.sleep(1)
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-            
-            # Assumes IX-like file structure
-            img_path = Path(volume, 'work', f'{platename}/{first_well}/img/{platename}_{first_well}.png')
-
-            # Wait for the image to be created
-            while not os.path.exists(img_path):
-                time.sleep(1)
-
-            # Checking the selection and changing the scale accordingly
-            if selection == 'motility':
-                scale = 'inferno'
-            else:
-                scale = 'gray'
-
-            # Open the image and create a figure
-            fig = create_figure_from_filepath(img_path, scale=scale)
-            
-            if 'Generating w1 thumbnails' in line:
-                # Return the command message, the figure, and the open status of the alerts
-                return command_message, fig, False, f'', False
-
-def cellprofile_wormsize_preview(wells, volume, platename, plate_base):
-    """
-    The purpose of this function is to preview the analysis of the selected options from
-    the configuration page, including worm size. This function will run wrmXpress and return
-    the path to the image, the figure, and the open status if the first well has not already been analyzed.
-    ===============================================================================
-    Arguments:
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-    ===============================================================================
-    Returns:
-        - path : str : Path to the image
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-        - open_status : bool : Open status of the alerts
-    ===============================================================================
-    """
-    first_well = wells[0]
-    print(first_well)
-    # Assumes IX-like file structure
-    first_well_path = Path(volume, f'output/straightened_worms/{plate_base}_{first_well}.tiff')
-
-    # Check if the first well path exists
-    if os.path.exists(first_well_path):
-        # Open the image and create a figure
-        fig = create_figure_from_filepath(first_well_path)
-
-        # Return the path and the figure and the open status of the alerts
-        return f"```{first_well_path}```", fig, False, f'', False
-
-    wrmxpress_command_split, output_preview_log_file, command_message, first_well = preamble_to_run_wrmXpress_preview(
-        platename=platename,
-        volume=volume,
-        wells=wells
-    )
-    with open(output_preview_log_file, 'w') as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-
-        while not os.path.exists(Path(volume, 'input')):
-            time.sleep(1)
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            if 'Generating w1 thumbnails' in line:
-                # Assumes IX-like file structure
-                file_path = Path(volume, f'output/straightened_worms/{plate_base}_{first_well}.tiff')
-                # Open the image and create a figure
-                fig = create_figure_from_filepath(file_path)
-
-                # Return the command message, the figure, and the open status of the alerts
-                return command_message, fig, False, f'', False
-
-def cellprofile_wormsize_intensity_cellpose_preview(wells, volume, platename, plate_base):
-    """
-    The purpose of this function is to preview the analysis of the selected options from
-    the configuration page, including worm size and intensity using CellPose. This function
-    will run wrmXpress and return the path to the image, the figure, and the open status if the first well has not already been analyzed.
-    ===============================================================================
-    Arguments:
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-    ===============================================================================
-    Returns:
-        - path : str : Path to the image
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-        - open_status : bool : Open status of the alerts
-    ===============================================================================
-    """
-    output_folder = Path(volume, 'input', platename)
-    output_preview_log_file = Path(output_folder, f'{platename}_preview.log')
-
-    # Assumes the first well in the list is the one to check
-    first_well = wells[0]
-    
-    # Check to see if first well already exists, if it does insert the img
-    # rather than running wrmXpress again
-    first_well_path = Path(volume, f'output/straightened_worms/{plate_base}_{first_well}.tiff')
-    if os.path.exists(first_well_path):
-        # Open the image and create a figure
-        fig = create_figure_from_filepath(first_well_path)
-
-        # Return the path and the figure and the open status of the alerts
-        return f"```{first_well_path}```", fig, False, f'', False
-
-    wrmxpress_command_split, output_preview_log_file, command_message, first_well = preamble_to_run_wrmXpress_preview(
-        platename=platename,
-        volume=volume,
-        wells=wells
-    )
-    with open(output_preview_log_file, 'w') as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            img_path = Path(volume, f'output/straightened_worms/{plate_base}_{first_well}.tiff')
-
-            # Wait for the image to be created
-            while not os.path.exists(img_path):
-                time.sleep(1)
-
-            # Open the image and create a figure
-            fig = create_figure_from_filepath(img_path)
-
-            if 'Generating w1 thumbnails' in line:
-                # Return the command message, the figure, and the open status of the alerts
-                return command_message, fig, False, f'', False
-            
-def cellprofile_mf_celltox_preview(wells, volume, platename, plate_base):
-    """
-    The purpose of this function is to preview the analysis of the selected options from
-    the configuration page, including motility, fecundity, and celltox. This function
-    will run wrmXpress and return the path to the image, the figure, and the open status if the first well has not already been analyzed.
-    ===============================================================================
-    Arguments:
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-    ===============================================================================
-    Returns:
-        - path : str : Path to the image
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-        - open_status : bool : Open status of the alerts
-    =============================================================================== 
-    """
-    output_folder = Path(volume, 'input', platename)
-    output_preview_log_file = Path(output_folder, f'{platename}_preview.log')
-
-    first_well = wells[0]  # Assuming first_well is defined here as the first item in wells
-
-    # Check to see if first well already exists, if it does insert the img
-    # rather than running wrmXpress again
-    first_well_path = Path(volume, 'work', platename, first_well, 'img', f'{platename}_{first_well}.png')
-    if os.path.exists(first_well_path):
-        # Open the image and create a figure
-        fig = create_figure_from_filepath(first_well_path)
-
-        # Return the path and the figure and the open status of the alerts
-        return f"```{first_well_path}```", fig, False, f'', False
-
-    wrmxpress_command_split, output_preview_log_file, command_message, first_well = preamble_to_run_wrmXpress_preview(
-        platename=platename,
-        volume=volume,
-        wells=wells
-    )
-    with open(output_preview_log_file, 'w') as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-
-        while not os.path.exists(output_folder):
-            time.sleep(1)
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            img_path = Path(volume, 'work', platename, first_well, 'img', f'{platename}_{first_well}.png')
-
-            # Wait for the image to be created
-            while not os.path.exists(img_path):
-                time.sleep(1)
-
-            # Open the image and create a figure
-            fig = create_figure_from_filepath(img_path)
-
-            if 'Generating w1 thumbnails' in line:
-                # Return the command message, the figure, and the open status of the alerts
-                return command_message, fig, False, f'', False
-
-def cellprofile_feeding_preview(wells, volume, platename, plate_base):
-    """
-    The purpose of this function is to preview the analysis of the selected options from
-    the configuration page, including feeding. This function will run wrmXpress and return
-    the path to the image, the figure, and the open status if the first well has not already been analyzed.
-    ===============================================================================
-    Arguments:
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-    ===============================================================================
-    Returns:
-        - path : str : Path to the image
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-        - open_status : bool : Open status of the alerts
-    ===============================================================================
-    """
-    output_folder = Path(volume, 'input', platename)
-    output_preview_log_file = Path(output_folder, f'{platename}_preview.log')
-
-    first_well = wells[0]  # Assuming first_well is the first item in wells list
-
-    # Check to see if first well already exists, if it does insert the img
-    # rather than running wrmXpress again
-    first_well_path = Path(volume, f'output/straightened_worms/{plate_base}-{first_well}.tiff')
-    if os.path.exists(first_well_path):
-        # Open the image and create a figure
-        fig = create_figure_from_filepath(first_well_path)
-
-        # Return the path and the figure and the open status of the alerts
-        return f"```{first_well_path}```", fig, False, f'', False
-
-    wrmxpress_command_split, output_preview_log_file, command_message, first_well = preamble_to_run_wrmXpress_preview(
-        platename=platename,
-        volume=volume,
-        wells=wells
-    )
-
-    with open(output_preview_log_file, 'w') as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-
-        while not os.path.exists(output_folder):
-            time.sleep(1)
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            if 'Generating w1 thumbnails' in line:
-                # Wait for the image to be created
-                while not os.path.exists(first_well_path):
-                    time.sleep(1)
-
-                # Open the image and create a figure
-                fig = create_figure_from_filepath(first_well_path)
-
-                # Return the command message, the figure, and the open status of the alerts
-                return command_message, fig, False, f'', False
-
-########################################################################
-####                                                                ####
-####                       app.py functions                         ####
-####                                                                ####
-########################################################################
-
-def preamble_to_run_wrmXpress_tracking(store):
-    """
-    The purpose of this function is to prepare the wrmXpress command, output folder, output file, 
-    command message, wells, volume, platename, motility, segment, cellprofiler, and cellprofilepipeline.
-    ===============================================================================
-    Arguments:
-        - store : dict : A dictionary containing the store
-    ===============================================================================
-    Returns:
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - command_message : str : A command message
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - motility : str : Motility
-        - segment : str : Segment
-        - cellprofiler : str : CellProfiler
-        - cellprofilepipeline : str : CellProfiler pipeline
-        - wells_analyzed : list : List of wells analyzed
-        - tracking_well : list : List of tracking wells
-    ===============================================================================
-
-    """
-    volume = store['mount']
-    platename = store['platename']
-    wells = store["wells"]
-    motility = store["motility"]
-    segment = store["segment"]
-    cellprofiler = store["cellprofiler"]
-    cellprofilepipeline = store["cellprofilepipeline"]
-    print('Running wrmXpress.')
-    # necessary file paths
-    img_dir = Path(volume, platename)
-    input_dir = Path(volume, 'input')
-    platename_input_dir = Path(input_dir, platename)
-    full_yaml = Path(volume, platename + '.yml')
-
-    update_yaml_file(
-            full_yaml,
-            full_yaml,
-            {'wells': ['All']}
-        )
-
-        # clean and create directories
-    clean_and_create_directories(
-            input_path=Path(volume, 'input', platename), 
-            work_path=Path(volume, 'work', platename),
-            output_path=Path(volume, 'output')
-        )
-
-    copy_files_to_input_directory(
-            platename_input_dir=platename_input_dir,
-            htd_file= None,
-            img_dir=img_dir,
-            wells=wells,
-            plate_base=None,
-            platename=platename
-        )
-
-        # Command message
-    command_message = f"```python wrmXpress/wrapper.py {platename}.yml {platename}```"
-
-    wrmxpress_command = f'python -u wrmXpress/wrapper.py {volume}/{platename}.yml {platename}'
-    wrmxpress_command_split = shlex.split(wrmxpress_command)
-    output_folder = Path(volume, 'work', platename)
-    output_file = Path(volume, 'work', platename, f"{platename}_run.log")  # Specify the name and location of the output file
-    wells_analyzed = []
-    tracking_well = []
-    return wrmxpress_command_split, output_folder, output_file, command_message, wells, volume, platename, motility, segment, cellprofiler, cellprofilepipeline, wells_analyzed, tracking_well
-
-def tracking_wrmXpress_run(
-    output_folder,
-    output_file,
-    wrmxpress_command_split,
-    volume,
-    platename,
-    wells,
-    wells_analyzed,
-    tracking_well,
-    set_progress
-    ):
-    """
-    The purpose of this function is to run wrmXpress for tracking and return the figure, open status, and command message.
-    ===============================================================================
-    Arguments:
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - wells : list : List of well names
-        - wells_analyzed : list : List of wells analyzed
-        - tracking_well : list : List of tracking wells
-        - set_progress : function : Function to set the progress
-    ===============================================================================
-    Returns:
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-    ===============================================================================
-    """
-    while not os.path.exists(output_folder):
-        time.sleep(1)
-    with open(output_file, "w") as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-
-        # Create an empty list to store the docker output
-        docker_output = []
-
-        for line in iter(process.stdout.readline, b''):
-            # Add the line to docker_output for further processing
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-            if 'Generating w1 thumbnails' in line:
-                tracks_file_path = Path(volume, 'output', 'thumbs', f'{platename}_tracks.png')
-
-                while not os.path.exists(tracks_file_path):
-                    time.sleep(1)
-
-                # create figure from file path
-                fig_1 = create_figure_from_filepath(tracks_file_path)
-                docker_output_formatted = ''.join(docker_output)
-
-                print('wrmXpress is finished')
-                docker_output.append('wrmXpress is finished')
-                docker_output_formatted = ''.join(docker_output)
-                return fig_1, False, False, f'', f'```{docker_output_formatted}```'
-            elif 'Reconfiguring' in line:
-                # find the well that is being analyzed
-                current_well = line.split('.')[0].split('_')[-1]
-
-                # add the well to the list of wells analyzed if it is not already there
-                if current_well not in wells_analyzed:
-                    wells_analyzed.append(current_well)
-                # obtain file path to current well
-                current_well_path = Path(volume, 'input', platename, 'TimePoint_1', f'{platename}_{wells_analyzed[-1]}.TIF')
-
-                # ensure file path exists
-                while not os.path.exists(current_well_path):
-                    time.sleep(1)
-
-                # create figure from file path
-                fig = create_figure_from_filepath(current_well_path)
-                docker_output_formatted = ''.join(docker_output)
-
-                set_progress((str(len(wells_analyzed)), str(2*len(wells)), fig, f'```{current_well_path}```' ,f'```{docker_output_formatted}```'))
-            
-            elif 'Tracking well' in line:
-                # find the well that is being analyzed
-                current_well = line.split(' ')[-1].split('.')[0]
-                # add the well to the list of wells analyzed if it is not already there
-                if current_well not in tracking_well:
-                    tracking_well.append(current_well)
-                
-                # obtain file path to current well
-                current_well_path = Path(volume, 'input', platename, 'TimePoint_1', f'{platename}_{tracking_well[-1]}.TIF')
-
-                # ensure file path exists
-                while not os.path.exists(current_well_path):
-                    time.sleep(1)
-                
-                # create figure from file path
-                fig = create_figure_from_filepath(current_well_path)
-                docker_output_formatted = ''.join(docker_output)
-
-                set_progress((str(len(tracking_well)+ len(wells_analyzed)), str(2*len(wells)), fig, f'```{current_well_path}```' ,f'```{docker_output_formatted}```'))
-
-def preamble_to_run_wrmXpress_non_tracking(store):
-    """
-    The purpose of this function is to prepare the wrmXpress command, output folder, output file,
-    command message, wells, volume, platename, motility, segment, cellprofiler, and cellprofilepipeline.
-    ===============================================================================
-    Arguments:
-        - store : dict : A dictionary containing the store
-    ===============================================================================
-    Returns:
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - command_message : str : A command message
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - motility : str : Motility
-        - segment : str : Segment
-        - cellprofiler : str : CellProfiler
-        - cellprofilepipeline : str : CellProfiler pipeline
-    ===============================================================================
-    """
-    volume = store['mount']
-    platename = store['platename']
-    wells = store["wells"]
-    motility = store["motility"]
-    segment = store["segment"]
-    cellprofiler = store["cellprofiler"]
-    cellprofilepipeline = store["cellprofilepipeline"]
-    print('Running wrmXpress.')
-
-    # necessary file paths
-    img_dir = Path(volume, platename)
-    input_dir = Path(volume, 'input')
-    platename_input_dir = Path(input_dir, platename)
-    plate_base = platename.split("_", 1)[0]
-    htd_file = Path(img_dir, f'{plate_base}.HTD')
-    full_yaml = Path(volume, platename + '.yml')
-
-    update_yaml_file(
-            full_yaml,
-            full_yaml,
-            {'wells': ['All']}
-        )
-
-        # clean and create directories
-    clean_and_create_directories(
-            input_path=Path(volume, 'input', platename), 
-            work_path=Path(volume, 'work', platename),
-            output_path=Path(volume, 'output')
-        )
-
-    copy_files_to_input_directory(
-            platename_input_dir=platename_input_dir,
-            htd_file=htd_file,
-            img_dir=img_dir,
-            plate_base=plate_base,
-            wells=wells,
-            platename=platename
-        )
-
-        # Command message
-    command_message = f"```python wrmXpress/wrapper.py {platename}.yml {platename}```"
-
-    wrmxpress_command = f'python -u wrmXpress/wrapper.py {volume}/{platename}.yml {platename}'
-    wrmxpress_command_split = shlex.split(wrmxpress_command)
-    output_folder = Path(volume, 'work', platename)
-    output_file = Path(volume, 'work', platename, f"{platename}_run.log")  # Specify the name and location of the output file
-    return wrmxpress_command_split, output_folder, output_file, command_message, wells, volume, platename, plate_base, motility, segment, cellprofiler, cellprofilepipeline
-
-def motility_or_segment_run(
-        output_folder,
-        output_file,
-        wrmxpress_command_split,
-        wells,
-        volume,
-        platename,
-        plate_base,
-        set_progress
-):
-    """
-    The purpose of this function is to run wrmXpress for motility or segment and return the figure, open status, and command message.
-    ===============================================================================
-    Arguments:
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-        - set_progress : function : Function to set the progress
-    ===============================================================================
-    Returns:
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-    ===============================================================================
-    """
-    while not os.path.exists(output_folder):
-                time.sleep(1)
-    with open(output_file, "w") as file:
-            
-        process = subprocess.Popen(
-                    wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-
-        # Create an empty list to store the docker output
-        docker_output = []
-        wells_analyzed = []
-        wells_to_be_analyzed = len(wells)
-
-        for line in iter(process.stdout.readline, b''):
-                    # Add the line to docker_output for further processing
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-            # Break the loop if 'Generating' is in the line
-            if "Generating" in line:
-                        break
-
-            # Process the line if 'Running' is in the line
-            if "Running" in line:
-                well_running = line.split(" ")[-1]
-                if well_running not in wells_analyzed:
-                    # Remove the '\n' from the well_running
-                    well_running = well_running.replace('\n', '')
-                    wells_analyzed.append(well_running)
-
-                    img_path = Path(
-                                volume, f'{platename}/TimePoint_1/{plate_base}_{wells_analyzed[-1]}.TIF')
-                    fig = create_figure_from_filepath(img_path)
-
-                    docker_output_formatted = ''.join(docker_output) 
-                    set_progress((
-                                str(len(wells_analyzed)),
-                                str(wells_to_be_analyzed),
-                                fig,
-                                f'```{img_path}```',
-                                f'```{docker_output_formatted}```'
-                            ))
-                
-
-        # get the platename (default) file in output dir that have .png extension
-        output_path = Path(volume, 'output', 'thumbs', platename + '.png')
-        while not os.path.exists(output_path):
-                    time.sleep(1)
-
-        # create a figure for the output
-        fig_1 = create_figure_from_filepath(output_path) 
-                
-        print('wrmXpress has finished.')
-        docker_output.append('wrmXpress has finished.')
-        docker_output_formatted = ''.join(docker_output) 
-                
-                    
-        # Return the figure, False, False, and an empty string
-        return fig_1, False, False, f'', f'```{docker_output_formatted}```'
-    
-def cellprofile_wormsize_run(output_folder, output_file, wrmxpress_command_split,
-                             wells, volume, platename, plate_base, set_progress,
-                             cellprofilepipeline):
-    """
-    The purpose of this function is to run wrmXpress for worm size and return the figure, open status, and command message.
-    ===============================================================================
-    Arguments:
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-        - set_progress : function : Function to set the progress
-        - cellprofilepipeline : str : CellProfiler pipeline
-    ===============================================================================
-    Returns:
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-    ===============================================================================
-    """
-    # Ensure the output folder exists
-    while not os.path.exists(output_folder):
-        time.sleep(1)
-    
-    # Open the output file and begin the subprocess
-    with open(output_file, "w") as file:
-        process = subprocess.Popen(wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-        wells_analyzed = []
-        wells_to_be_analyzed = len(wells)
-
-        # Read and process output from the subprocess
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            # Load the CSV file to find out which well is being analyzed
-            csv_file_path = Path(volume, 'input', f'image_paths_{cellprofilepipeline}.csv')
-            while not os.path.exists(csv_file_path):
-                time.sleep(1)
-            
-            read_csv = pd.read_csv(csv_file_path)
-            well_column = read_csv['Metadata_Well']  # Find the well column
-            
-            if 'Execution halted' in line:
-                output_path = Path(volume, 'output', 'thumbs', platename + '.png')
-                while not os.path.exists(output_path):
-                    time.sleep(1)
-
-                # Create a figure for the output
-                fig_1 = create_figure_from_filepath(output_path)
-                
-                print('wrmXpress has finished.')
-                docker_output.append('wrmXpress has finished.')
-                docker_output_formatted = ''.join(docker_output) 
-                return fig_1, False, False, '', f'```{docker_output_formatted}```'
-
-            # Check for the 'Image #' pattern to update progress
-            elif 'Image #' in line:
-                image_number_pattern = re.search(r'Image # (\d+)', line)
-                if image_number_pattern:
-                    image_number = int(image_number_pattern.group(1))
-                    well_id = well_column.iloc[image_number - 1]  # Adjust for zero indexing
-
-                    # Construct the image path
-                    img_path = Path(volume, f'input/{platename}/TimePoint_1/{plate_base}_{well_id}.TIF')
-                    if img_path.exists():
-                        # Load and display the image
-                        fig = create_figure_from_filepath(img_path)
-                        docker_output_formatted = ''.join(docker_output) 
-                        
-                        # Update progress
-                        set_progress((str(image_number), str(wells_to_be_analyzed), fig, f'```{img_path}```', f'```{docker_output_formatted}```'))
-
-def cellprofile_wormsize_intesity_cellpose_run(
-    output_folder, output_file, wrmxpress_command_split,
-    wells, volume, platename, plate_base, set_progress,
-    cellprofilepipeline
-):
-    """
-    The purpose of this function is to run wrmXpress for worm size and intensity using CellPose and return the figure, open status, and command message.
-    ===============================================================================
-    Arguments:
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-        - set_progress : function : Function to set the progress
-        - cellprofilepipeline : str : CellProfiler pipeline
-    ===============================================================================
-    Returns:
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-    ===============================================================================
-    """
-    while not os.path.exists(output_folder):
-        time.sleep(1)
-        
-    with open(output_file, "w") as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True
-        )
-        docker_output = []
-        wells_analyzed = []
-        wells_to_be_analyzed = len(wells)
-        progress = 0
-        total_progress = 2 * wells_to_be_analyzed
-        
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-            
-            if 'Generating w1 thumbnails' in line:
-                output_path = Path(volume, 'output', 'thumbs', platename + '.png')
-                while not os.path.exists(output_path):
-                    time.sleep(1)
-                
-                fig_1 = create_figure_from_filepath(output_path)
-                print('wrmXpress has finished.')
-                docker_output.append('wrmXpress has finished.')
-                docker_output_formatted = ''.join(docker_output)
-                return fig_1, False, False, f'', f'```{docker_output_formatted}```'
-            
-            elif 'Image #' in line:
-                csv_file_path = Path(volume, 'input', f'image_paths_{cellprofilepipeline}.csv')
-                while not os.path.exists(csv_file_path):
-                    time.sleep(1)
-                
-                read_csv = pd.read_csv(csv_file_path)
-                well_column = read_csv['Metadata_Well']
-                
-                image_number_pattern = re.search(r'Image # (\d+)', line)
-                if image_number_pattern:
-                    image_number = int(image_number_pattern.group(1))
-                    well_id = well_column.iloc[image_number - 1]
-                    img_path = Path(volume, f'input/{platename}/TimePoint_1/{plate_base}_{well_id}.TIF')
-                    if img_path.exists():
-                        fig = create_figure_from_filepath(img_path)
-                        progress += 1
-                        docker_output_formatted = ''.join(docker_output)
-                        set_progress((str(len(wells) + image_number), str(total_progress), fig, f'```{img_path}```', f'```{docker_output_formatted}```'))
-            
-            elif "[INFO]" in line and "%" in line:
-                info_parts = line.split('/')
-                info_well_analyzed = info_parts[0].split(' ')[-1]
-                info_total_wells = info_parts[1].split(' ')[0]
-                
-                if info_well_analyzed == info_total_wells:
-                    current_well = wells[int(info_well_analyzed)-1]
-                    img_path = Path(volume, f'input/{platename}/TimePoint_1/{plate_base}_{current_well}.TIF')
-                    if os.path.exists(img_path):
-                        fig = create_figure_from_filepath(img_path)
-                        docker_output_formatted = ''.join(docker_output)
-                        set_progress((str(progress), str(total_progress), fig, f'```{img_path}```', f'```{docker_output_formatted}```'))
-                else:
-                    current_well = wells[int(info_well_analyzed)]
-                    img_path = Path(volume, f'input/{platename}/TimePoint_1/{plate_base}_{current_well}.TIF')
-                    if os.path.exists(img_path):
-                        fig = create_figure_from_filepath(img_path, 'gray')
-                        docker_output_formatted = ''.join(docker_output)
-                        progress += 1
-                        set_progress((str(progress), str(total_progress), fig, f'```{img_path}```', f'```{docker_output_formatted}```'))
-
-def cellprofile_mf_celltox_run(output_folder, output_file, wrmxpress_command_split,
-                           wells, volume, platename, plate_base, set_progress,
-                           cellprofilepipeline):
-    """
-    The purpose of this function is to run wrmXpress for motility, fecundity, and celltox and return the figure, open status, and command message.
-    ===============================================================================
-    Arguments:
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-        - set_progress : function : Function to set the progress
-        - cellprofilepipeline : str : CellProfiler pipeline
-    ===============================================================================
-    Returns:
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-    ===============================================================================
-    """
-    while not os.path.exists(output_folder):
-        time.sleep(1)
-    
-    with open(output_file, "w") as file:
-        process = subprocess.Popen(wrmxpress_command_split, stdout=subprocess.PIPE,
-                                   stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-        wells_analyzed = []
-        wells_to_be_analyzed = len(wells)
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            csv_file_path = Path(volume, 'input', f'image_paths_{cellprofilepipeline}.csv')
-            while not os.path.exists(csv_file_path):
-                time.sleep(1)
-
-            read_csv = pd.read_csv(csv_file_path)
-            well_column = read_csv['Metadata_Well']
-            
-            if 'Generating w1 thumbnails' in line:
-                output_path = Path(volume, 'output', 'thumbs', platename + '.png')
-                while not os.path.exists(output_path):
-                    time.sleep(1)
-                
-                fig_1 = create_figure_from_filepath(output_path)
-                
-                print('wrmXpress has finished.')
-                docker_output.append('wrmXpress has finished.')
-                docker_output_formatted = ''.join(docker_output) 
-                return fig_1, False, False, '', f'```{docker_output_formatted}```'
-            
-            elif 'Image #' in line:
-                image_number_pattern = re.search(r'Image # (\d+)', line)
-                if image_number_pattern:
-                    image_number = int(image_number_pattern.group(1))
-                    well_id = well_column.iloc[image_number - 1]
-
-                    img_path = Path(volume, f'input/{platename}/TimePoint_1/{plate_base}_{well_id}_s1.TIF')
-                    if img_path.exists():
-                        fig = create_figure_from_filepath(img_path)
-                        docker_output_formatted = ''.join(docker_output) 
-                        set_progress((str(image_number), str(len(wells)), fig, f'```{img_path}```', f'```{docker_output_formatted}```'))
-
-def cellprofile_feeding_run(
-    output_folder, output_file, wrmxpress_command_split,
-    wells, volume, platename, plate_base, set_progress,
-    cellprofilepipeline
-):
-    """
-    The purpose of this function is to run wrmXpress for feeding and return the figure, open status, and command message.
-    ===============================================================================
-    Arguments:
-        - output_folder : str : Path to the output folder
-        - output_file : str : Path to the output file
-        - wrmxpress_command_split : list : List of wrmXpress commands
-        - wells : list : List of well names
-        - volume : str : Path to the volume
-        - platename : str : Name of the plate
-        - plate_base : str : Base name of the plate
-        - set_progress : function : Function to set the progress
-        - cellprofilepipeline : str : CellProfiler pipeline
-    ===============================================================================
-    Returns:
-        - fig : matplotlib.figure.Figure : A figure
-        - open_status : bool : Open status of the alerts
-        - command_message : str : A command message
-    ===============================================================================
-    """
-    while not os.path.exists(output_folder):
-        time.sleep(1)
-    
-    with open(output_file, "w") as file:
-        process = subprocess.Popen(
-            wrmxpress_command_split, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-        docker_output = []
-        wells_analyzed = []
-        wells_to_be_analyzed = len(wells)
-
-        for line in iter(process.stdout.readline, b''):
-            docker_output.append(line)
-            file.write(line)
-            file.flush()
-
-            csv_file_path = Path(volume, 'input', f'image_paths_{cellprofilepipeline}.csv')
-            while not os.path.exists(csv_file_path):
-                time.sleep(1)
-
-            read_csv = pd.read_csv(csv_file_path)
-            well_column = read_csv['Metadata_Well']
-            
-            if 'Generating w1 thumbnails' in line:
-                output_path = Path(volume, 'output', 'thumbs', platename + "_w1" + '.png')
-                while not os.path.exists(output_path):
-                    time.sleep(1)
-
-                fig_1 = create_figure_from_filepath(output_path)
-                print('wrmXpress has finished.')
-                docker_output.append('wrmXpress has finished.')
-                docker_output_formatted = ''.join(docker_output) 
-                return fig_1, False, False, '', f'```{docker_output_formatted}```'
-
-            elif 'Image #' in line:
-                image_number_pattern = re.search(r'Image # (\d+)', line)
-                if image_number_pattern:
-                    image_number = int(image_number_pattern.group(1))
-                    well_id = well_column.iloc[image_number - 1]
-
-                    img_path = Path(volume, f'input/{platename}/TimePoint_1/{plate_base}_{well_id}_w1.TIF')
-                    if img_path.exists():
-                        fig = create_figure_from_filepath(img_path)
-                        docker_output_formatted = ''.join(docker_output) 
-                        set_progress((str(image_number), str(len(wells)), fig, f'```{img_path}```', f'```{docker_output_formatted}```'))
 
 def formatting_module_for_yaml(pipeline):
     if pipeline == 'motility':
