+++ assets/custom.css
@@ -77,18 +77,18 @@ html .div-with-scroll::-webkit-scrollbar-thumb {
 }
 
 .accordion-button:not(.collapsed) {
-    font-size: 5rem;
+    font-size: 2rem;
     /* Adjust the size as needed to match h1 */
-    font-weight: bold;
+    /* font-weight: bold;
     /* h1 is usually bold */
 }
 
 .accordion-button {
-    font-size: 24px;
+    font-size: 1rem;
     /* Adjust as needed */
 }
 
 .custom-accordion-title {
-    font-size: 24px;
+    font-size: 1em;
     /* Adjust the size as needed */
 }
